Corrected map() function

+++ FlightComputerV2/include/core.h
@@ -1,1 +1,1 @@
-inline float map(double x, double in_min, double in_max, double out_min, double out_max) { return (double)(x - in_min) * (out_max - out_min) / (double)(in_max - in_min) + out_min; }
+inline double map(double x, double in_min, double in_max, double out_min, double out_max) { return (double)(x - in_min) * (out_max - out_min) / (double)(in_max - in_min) + out_min; }
